+++ src/actions/blend-tweens.js
@@ -1,6 +1,7 @@
-import Action from './';
+import tween from './tween';
+import { linear } from '../inc/easing';
 import { getProgressFromValue, getValueFromProgress } from '../inc/calc';
-import { clamp } from '../inc/transformers';
+import { flow, clamp, bezier } from '../inc/transformers';
 
 const clampProgress = clamp(0, 1);
 
@@ -9,95 +10,73 @@ function calcValueAtTime(from, to, duration, elapsed, ease) {
   return getValueFromProgress(from, to, progressAtTime);
 }
 
-class TweenBlend extends Action {
-  static defaultProps = {
-    accuracy: 60
-  };
+export default ({ from, to, onUpdate, accuracy = 60, ...props }) => {
+  // Get tween properties
+  const a = from;
+  const b = to;
+  const aCurrent = a.get();
+  const aDuration = a.getProp('duration');
+  const bDuration = b.getProp('duration');
+  const aEase = a.getProp('ease');
+  const bEase = b.getProp('ease');
+  const aFrom = a.getProp('from');
+  const bFrom = b.getProp('from');
+  const aTo = a.getProp('to');
+  const bTo = b.getProp('to');
 
-  onStart() {
-    const { from, to, duration, accuracy } = this.props;
-    const a = from;
-    const b = to;
-    const aDuration = a.getProp('duration');
-    const bDuration = b.getProp('duration');
+  // Analyse tweens
+  const overlapDuration = Math.min(aDuration - a.getElapsed(), bDuration);
+  const bValueAtTweenOverlapComplete = calcValueAtTime(bFrom, bTo, bDuration, overlapDuration, bEase);
+  const bStartsHigherThanA = (bFrom > aCurrent);
+  const bEndsHigherThanA = (bValueAtTweenOverlapComplete > aTo);
 
-    this.progress = 0;
-    this.duration = duration || Math.min(
-      aDuration - a.getElapsed(),
-      bDuration
-    );
+  // Blend points are defined as [t = time, v = valueAtTime]
+  // P1
+  // The start of the tween blend
+  const P1 = aCurrent;
+  const blendPoints = [P1];
 
-    const aEase = a.getProp('ease');
-    const bEase = b.getProp('ease');
-    const aFrom = a.getProp('from');
-    const bFrom = b.getProp('from');
-    const aTo = a.getProp('to');
-    const bTo = b.getProp('to');
-    const bValueAtBlendComplete = calcValueAtTime(bFrom, bTo, bDuration, this.duration, bEase);
-    const bStartsHigherThanA = bFrom > a.get();
-
-    this.blendPoints = [
-      [0, a.get()],
-      [this.duration, bValueAtBlendComplete]
-    ];
-
-    // Find tween intersection
-    const timestep = this.duration / accuracy;
-
-    let foundP1 = false;
-    let foundP2 = false;
+  // Find P2
+  // The possible intersection point between the two tweens
+  const tweensIntersect = (bStartsHigherThanA !== bEndsHigherThanA);
+  if (tweensIntersect) {
+    let timestep = overlapDuration / accuracy;
 
     for (let i = 0; i < accuracy; i++) {
-      const totalTime = i * timestep;
+      const runningTime = i * timestep;
 
-      const aValueAtTime = calcValueAtTime(aFrom, aTo, aDuration, a.elapsed + totalTime, aEase);
-      const bValueAtTime = calcValueAtTime(bFrom, bTo, bDuration, b.elapsed + totalTime, bEase);
+      const aValueAtTime = calcValueAtTime(aFrom, aTo, aDuration, a.elapsed + runningTime, aEase);
+      const bValueAtTime = calcValueAtTime(bFrom, bTo, bDuration, b.elapsed + runningTime, bEase);
 
       const hasIntersected = (
         (bStartsHigherThanA && aValueAtTime > bValueAtTime) ||
         (!bStartsHigherThanA && aValueAtTime < bValueAtTime)
       );
 
-      if (!foundP1 && hasIntersected) {
-        this.blendPoints.splice(1, 0, [totalTime, bValueAtTime]);
-        foundP1 = true;
+      if (hasIntersected) {
+        const P2 = bValueAtTime;
+        blendPoints.push(P2);
+        break;
       }
-
-      const hasIntersectedB = (
-        (bStartsHigherThanA && aValueAtTime < bValueAtTime) ||
-        (!bStartsHigherThanA && aValueAtTime > bValueAtTime)
-      );
-
-      if (foundP1 && hasIntersectedB) {
-        this.blendPoints[2] = [totalTime, bValueAtTime];
-        foundP2 = true;
-      }
-
-      if (foundP2) break;
-    }
-
-    to.start();
-  }
-
-  update() {
-    const { to } = this.props;
-    const b = to;
-
-    this.progress = clampProgress(getProgressFromValue(this.blendPoints[0][0], this.blendPoints[2][0], b.getElapsed()));
-
-    if (this.progress >= 1) {
-      return b.get();
-    } else {
-      const aP = getValueFromProgress(this.blendPoints[0][1], this.blendPoints[1][1], this.progress);
-      const bP = getValueFromProgress(this.blendPoints[1][1], this.blendPoints[2][1], this.progress);
-      return getValueFromProgress(aP, bP, this.progress);
     }
   }
 
-  isActionComplete() {
-    const { to } = this.props;
-    return !to.isActive();
-  }
-}
+  // P3
+  // The points the two tweens stop overlapping
+  const P3 = bValueAtTweenOverlapComplete;
+  blendPoints.push(P3);
 
-export default (props) => new TweenBlend(props);
+  // Push the end state of b tween as final control point
+  const P4 = bTo;
+  blendPoints.push(P4);
+
+  return tween({
+    ...props,
+    duration: bDuration,
+    ease: linear,
+    onUpdate: flow(
+      bezier(blendPoints),
+      onUpdate
+    )
+  });
+};